+++ Setup-Your-Mac-via-Dialog.bash
@@ -9,18 +9,15 @@
 #
 # HISTORY
 #
-#   Version 1.8.0, 06-Mar-2023, Dan K. Snelson (@dan-snelson)
-#   - Introduces "Configurations" (thanks, @drtaru!)
-#       - Required
-#       - Recommended
-#       - Complete
-#   - Play video at Welcome dialog (Issue No. 36)
-#
-#   Version 1.8.1, 11-Mar-2023, Dan K. Snelson (@dan-snelson)
-#   - Added `currentLoggedInUser` function to better validate `loggedInUser` (Issue No. 2)
-#   - Added new "Microsoft Office 365" Remote Validation (Pull Request No. 3)
-#   - Improved logging when `welcomeDialog` is `video` or `false` (Issue No. 4)
-#   - Create `overlayicon` from Self Service's custom icon (thanks, Mike Schwartz!)
+#   Version 1.9.0, 01-Apr-2023, Dan K. Snelson (@dan-snelson)
+#   - Previously installed apps with a `filepath` validation now display "Previously Installed" (instead of a generic "Installed"; Issue No. 13; thanks for the idea, @Manikandan!)
+#   - Allow "first name" to correctly handle names in "Lastname, Firstname" format (Pull Request No. 11; thanks, @meschwartz!)
+#   - Corrected `PATH` (thanks, @Theile!)
+#   - `Configuration` no longer displays in SYM's `infobox` when `welcomeDialog` is set to `false` or `video` (Issue No. 12; thanks, @Manikandan!)
+#   - Updated icon hashes
+#   - New `toggleJamfLaunchDaemon` function (Pull Request No. 16; thanks, @robjschroeder!)
+#   - Formatted policyJSON with [Erik Lynd's JSON Tools](https://marketplace.visualstudio.com/items?itemName=eriklynd.json-tools)
+#   - Corrected an issue where inventory would be submitted twice (thanks, @Manikandan!)
 #
 ####################################################################################################
 
@@ -36,13 +33,13 @@
 # Script Version and Jamf Pro Script Parameters
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-scriptVersion="1.8.1"
-export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/bin/
+scriptVersion="1.9.0"
+export PATH=/usr/bin:/bin:/usr/sbin:/sbin
 scriptLog="${4:-"/var/log/org.churchofjesuschrist.log"}"                        # Parameter 4: Script Log Location [ /var/log/org.churchofjesuschrist.log ] (i.e., Your organization's default location for client-side logs)
 debugMode="${5:-"verbose"}"                                                     # Parameter 5: Debug Mode [ verbose (default) | true | false ]
 welcomeDialog="${6:-"userInput"}"                                               # Parameter 6: Welcome dialog [ userInput (default) | video | false ]
 completionActionOption="${7:-"Restart Attended"}"                               # Parameter 7: Completion Action [ wait | sleep (with seconds) | Shut Down | Shut Down Attended | Shut Down Confirm | Restart | Restart Attended (default) | Restart Confirm | Log Out | Log Out Attended | Log Out Confirm ]
-requiredMinimumBuild="${8:-"disabled"}"                                         # Parameter 8: Required Minimum Build [ disabled (default) | 22D ] (i.e., Your organization's required minimum build of macOS to allow users to proceed; use "22D" for macOS 13.2.x)
+requiredMinimumBuild="${8:-"disabled"}"                                         # Parameter 8: Required Minimum Build [ disabled (default) | 22E ] (i.e., Your organization's required minimum build of macOS to allow users to proceed; use "22E" for macOS 13.3)
 outdatedOsAction="${9:-"/System/Library/CoreServices/Software Update.app"}"     # Parameter 9: Outdated OS Action [ /System/Library/CoreServices/Software Update.app (default) | jamfselfservice://content?entity=policy&id=117&action=view ] (i.e., Jamf Pro Self Service policy ID for operating system ugprades)
 
 
@@ -209,15 +206,15 @@ counter="1"
 
 until { [[ "${loggedInUser}" != "_mbsetupuser" ]] || [[ "${counter}" -gt "180" ]]; } && { [[ "${loggedInUser}" != "loginwindow" ]] || [[ "${counter}" -gt "30" ]]; } ; do
 
-	updateScriptLog "PRE-FLIGHT CHECK: Logged-in User Counter: ${counter}"
-	currentLoggedInUser
-	sleep 2
-	((counter++))
+    updateScriptLog "PRE-FLIGHT CHECK: Logged-in User Counter: ${counter}"
+    currentLoggedInUser
+    sleep 2
+    ((counter++))
 
 done
 
 loggedInUserFullname=$( id -F "${loggedInUser}" )
-loggedInUserFirstname=$( echo "$loggedInUserFullname" | cut -d " " -f 1 )
+loggedInUserFirstname=$( echo "$loggedInUserFullname" | sed -E 's/^.*, // ; s/([^ ]*).*/\1/' )
 loggedInUserID=$( id -u "${loggedInUser}" )
 updateScriptLog "PRE-FLIGHT CHECK: Current Logged-in User First Name: ${loggedInUserFirstname}"
 updateScriptLog "PRE-FLIGHT CHECK: Current Logged-in User ID: ${loggedInUserID}"
@@ -225,19 +222,44 @@ updateScriptLog "PRE-FLIGHT CHECK: Current Logged-in User ID: ${loggedInUserID}"
 
 
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
-# Pre-flight Check: Temporarily disable `jamf` binary check-in (thanks, @mactroll and @cube!)
+# Pre-flight Check: Toggle `jamf` binary check-in (thanks, @robjschroeder!)
+# shellcheck disable=SC2143
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
+function toggleJamfLaunchDaemon() {
+    
+jamflaunchDaemon="/Library/LaunchDaemons/com.jamfsoftware.task.1.plist"
 if [[ "${debugMode}" == "true" ]] || [[ "${debugMode}" == "verbose" ]] ; then
-    updateScriptLog "PRE-FLIGHT CHECK: DEBUG MODE: Normally, 'jamf' binary check-in would be temporarily disabled"
+    if [[ $(/bin/launchctl list | grep com.jamfsoftware.task.E) ]]; then
+        updateScriptLog "PRE-FLIGHT CHECK: DEBUG MODE: Normally, 'jamf' binary check-in would be temporarily disabled"
+    else
+        updateScriptLog "QUIT SCRIPT: DEBUG MODE: Normally, 'jamf' binary check-in would be re-enabled"
+    fi
 else
-    updateScriptLog "PRE-FLIGHT CHECK: Temporarily disable 'jamf' binary check-in"
-    jamflaunchDaemon="/Library/LaunchDaemons/com.jamfsoftware.task.1.plist"
     while [[ ! -f "${jamflaunchDaemon}" ]] ; do
         sleep 0.1
     done
-    /bin/launchctl bootout system "$jamflaunchDaemon"
+    if [[ $(/bin/launchctl list | grep com.jamfsoftware.task.E) ]]; then
+        updateScriptLog "PRE-FLIGHT CHECK: Temporarily disable 'jamf' binary check-in"
+        /bin/launchctl bootout system "${jamflaunchDaemon}"
+    else
+        updateScriptLog "QUIT SCRIPT: Re-enabling 'jamf' binary check-in"
+        updateScriptLog "QUIT SCRIPT: 'jamf' binary check-in daemon not loaded, attempting to bootstrap and start"
+        result="0"
+        until [ $result -eq 3 ]; do
+            /bin/launchctl bootstrap system "${jamflaunchDaemon}" && /bin/launchctl start "${jamflaunchDaemon}"
+            result="$?"
+            if [ $result = 3 ]; then
+                updateScriptLog "QUIT SCRIPT: Staring 'jamf' binary check-in daemon"
+            else
+                updateScriptLog "QUIT SCRIPT: Failed to start 'jamf' binary check-in daemon"
+            fi
+        done
+    fi
 fi
+}
+
+toggleJamfLaunchDaemon
 
 
 
@@ -294,7 +316,7 @@ function dialogCheck() {
 
     else
 
-        updateScriptLog "PRE-FLIGHT CHECK: swiftDialog version $(dialog --version) found; proceeding..."
+        updateScriptLog "PRE-FLIGHT CHECK: swiftDialog version $(/usr/local/bin/dialog --version) found; proceeding..."
 
     fi
 
@@ -302,6 +324,8 @@ function dialogCheck() {
 
 if [[ ! -e "/Library/Application Support/Dialog/Dialog.app" ]]; then
     dialogCheck
+else
+    updateScriptLog "PRE-FLIGHT CHECK: swiftDialog version $(/usr/local/bin/dialog --version) found; proceeding..."
 fi
 
 
@@ -371,7 +395,7 @@ welcomeMessage="Please enter your Mac's 5-Digit **Asset Tag** Number located und
 welcomeBannerImage="https://img.freepik.com/free-vector/abstract-blue-geometric-polygonal-background_1035-18996.jpg?w=1060&t=st=1678480429~exp=1678481029~hmac=92269a18199171b33b9a2954161b459c89735ec3fefb8a8475365b6b007e328e"
 welcomeBannerText="Welcome to your new Mac, ${loggedInUserFirstname}!"
 welcomeCaption="Please review the above video, then click Continue."
-welcomeVideoID="vimeoid=803933536"
+welcomeVideoID="vimeoid=812753953"
 
 # Welcome icon set to either light or dark, based on user's Apperance setting (thanks, @mm2270!)
 appleInterfaceStyle=$( /usr/bin/defaults read /Users/"${loggedInUser}"/Library/Preferences/.GlobalPreferences.plist AppleInterfaceStyle 2>&1 )
@@ -405,7 +429,8 @@ welcomeVideo="--title \"$welcomeTitle\" \
 # "Welcome" JSON for Capturing User Input (thanks, @bartreardon!)
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-welcomeJSON='{
+welcomeJSON='
+{
     "bannerimage" : "'"${welcomeBannerImage}"'",
     "bannertext" : "'"${welcomeBannerText}"'",
     "title" : "'"${welcomeTitle}"'",
@@ -417,7 +442,7 @@ welcomeJSON='{
     "infotext" : "'"${scriptVersion}"'",
     "blurscreen" : "true",
     "ontop" : "true",
-    "titlefont" : "shadow=true, size=40",
+    "titlefont" : "shadow=true, size=28",
     "messagefont" : "size=14",
     "textfield" : [
         {   "title" : "Computer Name",
@@ -435,7 +460,7 @@ welcomeJSON='{
             "regexerror" : "Please enter five digits for the Asset Tag,"
         }
     ],
-  "selectitems" : [
+    "selectitems" : [
         {   "title" : "Configuration",
             "default" : "Required",
             "values" : [
@@ -477,7 +502,8 @@ welcomeJSON='{
         }
     ],
     "height" : "725"
-}'
+}
+'
 
 
 
@@ -495,10 +521,10 @@ title="Setting up ${loggedInUserFirstname}'s Mac"
 message="Please wait while the following apps are installed …"
 bannerImage="https://img.freepik.com/free-photo/yellow-watercolor-paper_95678-446.jpg"
 bannerText="Setting up ${loggedInUserFirstname}'s Mac"
-helpmessage="If you need assistance, please contact the Global Service Department:  \n- **Telephone:** +1 (801) 555-1212  \n- **Email:** support@domain.org  \n- **Knowledge Base Article:** KB0057050  \n\n**Computer Information:** \n\n- **Operating System:**  ${macOSproductVersion} ($macOSbuildVersion)  \n- **Serial Number:** ${serialNumber}  \n- **Dialog:** ${dialogVersion}  \n- **Started:** ${timestamp}"
+helpmessage="If you need assistance, please contact the Global Service Department:  \n- **Telephone:** +1 (801) 555-1212  \n- **Email:** support@domain.org  \n- **Knowledge Base Article:** KB0057050  \n\n**Computer Information:**  \n- **Operating System:**  ${macOSproductVersion} (${macOSbuildVersion})  \n- **Serial Number:** ${serialNumber}  \n- **Dialog:** ${dialogVersion}  \n- **Started:** ${timestamp}"
 infobox="Analyzing input …" # Customize at "Update Setup Your Mac's infobox"
 
-# Create `overlayicon` from Self Service's custom icon (thanks, Mike Schwartz!)
+# Create `overlayicon` from Self Service's custom icon (thanks, @meschwartz!)
 xxd -p -s 260 "$(defaults read /Library/Preferences/com.jamfsoftware.jamf self_service_app_path)"/Icon$'\r'/..namedfork/rsrc | xxd -r -p > /var/tmp/overlayicon.icns
 overlayicon="/var/tmp/overlayicon.icns"
 
@@ -562,14 +588,14 @@ dialogSetupYourMacCMD="$dialogBinary \
 #
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 #
-# From @wakco: If you would prefer to get your policyJSON externally replace it with:
+# Thanks, @wakco: If you would prefer to get your policyJSON externally replace it with:
 #  - policyJSON="$(cat /path/to/file.json)" # For getting from a file, replacing /path/to/file.json with the path to your file, or
 #  - policyJSON="$(curl -sL https://server.name/jsonquery)" # For a URL, replacing https://server.name/jsonquery with the URL of your file.
 #
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 #
-# From @astrugatch: I added this line to global variables:
-# jsonURL=${10}                                                               # URL Hosting JSON for policy_array
+# Thanks, @astrugatch: I added this line to global variables:
+# jsonURL=${10} # URL Hosting JSON for policy_array
 #
 # And this line replaces the entirety of the policy_array (~ line 503):
 # policy_array=("$(curl -sL $jsonURL)")
@@ -677,7 +703,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Final Configuration",
-                        "icon": "00d7c19b984222630f20b6821425c3548e4b5094ecd846b03bde0994aaf08826",
+                        "icon": "4723e3e341a7e11e6881e418cf91b157fcc081bdb8948697750e5da3562df728",
                         "progresstext": "Finalizing Configuration …",
                         "trigger_list": [
                             {
@@ -692,7 +718,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Computer Inventory",
-                        "icon": "90958d0e1f8f8287a86a1198d21cded84eeea44886df2b3357d909fe2e6f1296",
+                        "icon": "ff2147a6c09f5ef73d1c4406d00346811a9c64c0b6b7f36eb52fcb44943d26f9",
                         "progresstext": "A listing of your Mac’s apps and settings — its inventory — is sent automatically to the Jamf Pro server daily.",
                         "trigger_list": [
                             {
@@ -809,7 +835,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Final Configuration",
-                        "icon": "00d7c19b984222630f20b6821425c3548e4b5094ecd846b03bde0994aaf08826",
+                        "icon": "4723e3e341a7e11e6881e418cf91b157fcc081bdb8948697750e5da3562df728",
                         "progresstext": "Finalizing Configuration …",
                         "trigger_list": [
                             {
@@ -824,7 +850,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Computer Inventory",
-                        "icon": "90958d0e1f8f8287a86a1198d21cded84eeea44886df2b3357d909fe2e6f1296",
+                        "icon": "ff2147a6c09f5ef73d1c4406d00346811a9c64c0b6b7f36eb52fcb44943d26f9",
                         "progresstext": "A listing of your Mac’s apps and settings — its inventory — is sent automatically to the Jamf Pro server daily.",
                         "trigger_list": [
                             {
@@ -960,7 +986,8 @@ function policyJSONConfiguration() {
                                 "validation": "/Applications/Microsoft Teams.app/Contents/Info.plist"
                             }
                         ]
-                    },                    {
+                    },
+                    {
                         "listitem": "Adobe Acrobat Reader",
                         "icon": "988b669ca27eab93a9bcd53bb7e2873fb98be4eaa95ae8974c14d611bea1d95f",
                         "progresstext": "Views, prints, and comments on PDF documents, and connects to Adobe Document Cloud.",
@@ -984,7 +1011,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Final Configuration",
-                        "icon": "00d7c19b984222630f20b6821425c3548e4b5094ecd846b03bde0994aaf08826",
+                        "icon": "4723e3e341a7e11e6881e418cf91b157fcc081bdb8948697750e5da3562df728",
                         "progresstext": "Finalizing Configuration …",
                         "trigger_list": [
                             {
@@ -999,7 +1026,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Computer Inventory",
-                        "icon": "90958d0e1f8f8287a86a1198d21cded84eeea44886df2b3357d909fe2e6f1296",
+                        "icon": "ff2147a6c09f5ef73d1c4406d00346811a9c64c0b6b7f36eb52fcb44943d26f9",
                         "progresstext": "A listing of your Mac’s apps and settings — its inventory — is sent automatically to the Jamf Pro server daily.",
                         "trigger_list": [
                             {
@@ -1076,7 +1103,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Final Configuration",
-                        "icon": "00d7c19b984222630f20b6821425c3548e4b5094ecd846b03bde0994aaf08826",
+                        "icon": "4723e3e341a7e11e6881e418cf91b157fcc081bdb8948697750e5da3562df728",
                         "progresstext": "Finalizing Configuration …",
                         "trigger_list": [
                             {
@@ -1091,7 +1118,7 @@ function policyJSONConfiguration() {
                     },
                     {
                         "listitem": "Computer Inventory",
-                        "icon": "90958d0e1f8f8287a86a1198d21cded84eeea44886df2b3357d909fe2e6f1296",
+                        "icon": "ff2147a6c09f5ef73d1c4406d00346811a9c64c0b6b7f36eb52fcb44943d26f9",
                         "progresstext": "A listing of your Mac’s apps and settings — its inventory — is sent automatically to the Jamf Pro server daily.",
                         "trigger_list": [
                             {
@@ -1371,8 +1398,8 @@ function run_jamf_trigger() {
     elif [[ "$trigger" == "recon" ]]; then
 
         dialogUpdateSetupYourMac "listitem: index: $i, status: wait, statustext: Updating …, "
-        updateScriptLog "SETUP YOUR MAC DIALOG: Updating computer inventory with the following reconOptions: \"${reconOptions}\" …"
-        eval "${jamfBinary} recon ${reconOptions}"
+        updateScriptLog "SETUP YOUR MAC DIALOG: Computer inventory, with the following reconOptions: \"${reconOptions}\", will be be executed in the 'confirmPolicyExecution' function …"
+        # eval "${jamfBinary} recon ${reconOptions}"
 
     else
 
@@ -1407,8 +1434,10 @@ function confirmPolicyExecution() {
                 sleep 1
             elif [[ -f "${validation}" ]]; then
                 updateScriptLog "SETUP YOUR MAC DIALOG: Confirm Policy Execution: ${validation} exists; skipping 'run_jamf_trigger ${trigger}'"
+                previouslyInstalled="true"
             else
                 updateScriptLog "SETUP YOUR MAC DIALOG: Confirm Policy Execution: ${validation} does NOT exist; executing 'run_jamf_trigger ${trigger}'"
+                previouslyInstalled="false"
                 run_jamf_trigger "${trigger}"
             fi
             ;;
@@ -1459,7 +1488,9 @@ function validatePolicyResult() {
 
         */* ) 
             updateScriptLog "SETUP YOUR MAC DIALOG: Validate Policy Result: Testing for \"$validation\" …"
-            if [[ -f "${validation}" ]]; then
+            if [[ "${previouslyInstalled}" == "true" ]]; then
+                dialogUpdateSetupYourMac "listitem: index: $i, status: success, statustext: Previously Installed"
+            elif [[ -f "${validation}" ]]; then
                 dialogUpdateSetupYourMac "listitem: index: $i, status: success, statustext: Installed"
             else
                 dialogUpdateSetupYourMac "listitem: index: $i, status: fail, statustext: Failed"
@@ -1498,9 +1529,9 @@ function validatePolicyResult() {
                             jamfProPolicyNameFailures+="• $listitem  \n"
                         fi
                     else
-                        # Inelligible
+                        # Ineligible
                         updateScriptLog "SETUP YOUR MAC DIALOG: Locally Validate Policy Result: Rosetta 2 is not applicable"
-                        dialogUpdateSetupYourMac "listitem: index: $i, status: error, statustext: Inelligible"
+                        dialogUpdateSetupYourMac "listitem: index: $i, status: error, statustext: Ineligible"
                     fi
                     ;;
                 filevault )
@@ -1845,10 +1876,8 @@ function quitScript() {
     updateScriptLog "QUIT SCRIPT: De-caffeinate …"
     killProcess "caffeinate"
 
-    # Reenable 'jamf' binary check-in
-    # Purposely commented-out on 2023-01-26-092705; presumes Mac will be rebooted
-    # updateScriptLog "QUIT SCRIPT: Reenable 'jamf' binary check-in"
-    # launchctl bootstrap system "${jamflaunchDaemon}"
+    # Toggle `jamf` binary check-in 
+    toggleJamfLaunchDaemon
 
     # Remove overlayicon
     if [[ -e ${overlayicon} ]]; then
@@ -2215,13 +2244,20 @@ dialogUpdateWelcome "quit:"
 # Output Line Number in `verbose` Debug Mode
 if [[ "${debugMode}" == "verbose" ]]; then updateScriptLog "# # # SETUP YOUR MAC VERBOSE DEBUG MODE: Line No. ${LINENO} # # #" ; fi
 
+# When `welcomeDialog` is set to `false` or `video`, set the value of `infoboxConfiguration` to null (thanks for the idea, @Manikandan!)
+if [[ "${symConfiguration}" == *"Catch-all"* ]]; then
+    infoboxConfiguration=""
+else
+    infoboxConfiguration="${symConfiguration}"
+fi
+
 infobox=""
 
 if [[ -n ${comment} ]]; then infobox+="**Comment:**  \n$comment  \n\n" ; fi
 if [[ -n ${computerName} ]]; then infobox+="**Computer Name:**  \n$computerName  \n\n" ; fi
 if [[ -n ${userName} ]]; then infobox+="**Username:**  \n$userName  \n\n" ; fi
 if [[ -n ${assetTag} ]]; then infobox+="**Asset Tag:**  \n$assetTag  \n\n" ; fi
-if [[ -n ${symConfiguration} ]]; then infobox+="**Configuration:**  \n$symConfiguration  \n\n" ; fi
+if [[ -n ${infoboxConfiguration} ]]; then infobox+="**Configuration:**  \n$infoboxConfiguration  \n\n" ; fi
 if [[ -n ${department} ]]; then infobox+="**Department:**  \n$department  \n\n" ; fi
 
 dialogUpdateSetupYourMac "infobox: ${infobox}"